fix(core): add missing flags and fix imports
